performed calculations in the main using the calculations method and fixed expected output for integer division to accomodate the potential double data type return

+++ CalculatorObject/src/main/java/com/company/Calculations.java
@@ -20,8 +20,8 @@ public class Calculations {
         return product;
     }
 
-    public int divide(int a, int b) {
-        int quotient;
+    public double divide(int a, int b) {
+        double quotient;
         quotient = a / b;
         return quotient;
     }
